skills/avatar-renderer/runtime/pyproject.toml
[project]
name = "avatar-renderer"
version = "0.1.0"
description = "Render avatar images with emotional facial expressions and animation sequences."
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pyright]
pythonVersion = "3.10"
include = ["."]
typeCheckingMode = "standard"
